+++ src/index.ts
@@ -3,6 +3,10 @@ export { Application } from "./lib/application";
 export { EventDispatcher, Event } from "./lib/utils/events";
 export { resetReflectionID } from "./lib/models/reflections/abstract";
 export { normalizePath } from "./lib/utils/fs";
+/**
+ * All symbols documented under the Models namespace are also available in the root import.
+ */
+export * as Models from "./lib/models";
 export * from "./lib/models";
 export {
     Converter,