+++ src/lib/output/themes/default/assets/typedoc/Application.ts
@@ -37,6 +37,7 @@ export class Application {
      */
     constructor() {
         this.createComponents(document.body);
+        this.ensureActivePageVisible();
         this.ensureFocusedElementVisible();
         window.addEventListener("hashchange", () =>
             this.ensureFocusedElementVisible()
@@ -61,6 +62,21 @@ export class Application {
         this.ensureFocusedElementVisible();
     }
 
+    private ensureActivePageVisible() {
+        const pageLink = document.querySelector(".tsd-navigation .current");
+        let iter = pageLink?.parentElement;
+        while (iter && !iter.classList.contains(".tsd-navigation")) {
+            // Expand parent namespaces if collapsed, don't expand current namespace
+            if (
+                iter instanceof HTMLDetailsElement &&
+                pageLink?.parentElement?.parentElement !== iter
+            ) {
+                iter.open = true;
+            }
+            iter = iter.parentElement;
+        }
+    }
+
     /**
      * Ensures that if a user was linked to a reflection which is hidden because of filter
      * settings, that reflection is still shown.
@@ -71,6 +87,8 @@ export class Application {
             this.alwaysVisibleMember.firstElementChild!.remove();
             this.alwaysVisibleMember = null;
         }
+
+        if (!location.hash) return;
 
         const reflAnchor = document.getElementById(location.hash.substring(1));
         if (!reflAnchor) return;